debugging started

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: samcasti <samcasti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 12:52:07 by samcasti          #+#    #+#             */
-/*   Updated: 2024/04/22 13:42:18 by samcasti         ###   ########.fr       */
+/*   Updated: 2024/04/25 12:38:01 by samcasti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
